requirements.txt
requests==2.24.0
facebook-sdk==3.1.0
python-dotenv==0.15.0
re==2.2.1
PIL==8.0.1